// src/order/domain/enums/order-status.enum.ts
export enum OrderStatus {
  RECEIVED = "RECEIVED",
  IN_DIAGNOSIS = "IN_DIAGNOSIS",
  WAITING_BUDGET_APPROVAL = "WAITING_BUDGET_APPROVAL",
  BUDGET_APPROVED = "BUDGET_APPROVED",
  BUDGET_REJECTED = "BUDGET_REJECTED",
  IN_EXECUTION = "IN_EXECUTION",
  FINISHED = "FINISHED",
  DELIVERED = "DELIVERED",
  CANCELLED = "CANCELLED",
}